# openapi.json
{"openapi":"3.0.2","info":{"title":"movies","version":"0.1.0"},"paths":{"/api/v1/films":{"get":{"tags":["films"],"summary":"Главная страница фильмов.","description":"На ней выводятся популярные фильмы, с указанием поля сортировки и жанра.","operationId":"film_list_api_v1_films_get","parameters":[{"description":"Сортировка по указанному полю.","required":false,"schema":{"title":"Sort","anyOf":[{"$ref":"#/components/schemas/imbdbOrderEnum"},{"$ref":"#/components/schemas/idPersonOrderEnum"}],"description":"Сортировка по указанному полю."},"name":"sort","in":"query"},{"description":"Номер страницы.","required":false,"schema":{"title":"Page[Number]","type":"integer","description":"Номер страницы.","default":0,"gte":0},"name":"page[number]","in":"query"},{"description":"Количество элементов на странице.","required":false,"schema":{"title":"Page[Size]","exclusiveMinimum":0.0,"type":"integer","description":"Количество элементов на странице.","default":25,"lte":50},"name":"page[size]","in":"query"},{"description":"Текст полнотекстового поиска.","required":false,"schema":{"title":"Query","type":"string","description":"Текст полнотекстового поиска."},"name":"query","in":"query"},{"description":"Список id's фильмов, для фильтрации.","required":false,"schema":{"title":"Ids","type":"array","items":{"type":"string","format":"uuid"},"description":"Список id's фильмов, для фильтрации."},"name":"ids","in":"query"},{"description":"Фильтр по id жанра.","required":false,"schema":{"title":"Filter[Genre]","type":"string","description":"Фильтр по id жанра.","format":"uuid"},"name":"filter[genre]","in":"query"},{"description":"Фильтр по названию жанра.","required":false,"schema":{"title":"Filter[Genre.Name]","type":"string","description":"Фильтр по названию жанра."},"name":"filter[genre.name]","in":"query"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"film not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"films not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Список фильмов.","content":{"application/json":{"schema":{"title":"Response Film List Api V1 Films Get","type":"array","items":{"$ref":"#/components/schemas/Film"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/search":{"get":{"tags":["films"],"summary":"Полнотекстовый поиск по фильмам.","description":"Полнотекстовый поиск фильма, по указанному параметру query в запросе.","operationId":"film_list_search_api_v1_films_search_get","parameters":[{"description":"Сортировка по указанному полю.","required":false,"schema":{"title":"Sort","anyOf":[{"$ref":"#/components/schemas/imbdbOrderEnum"},{"$ref":"#/components/schemas/idPersonOrderEnum"}],"description":"Сортировка по указанному полю."},"name":"sort","in":"query"},{"description":"Номер страницы.","required":false,"schema":{"title":"Page[Number]","type":"integer","description":"Номер страницы.","default":0,"gte":0},"name":"page[number]","in":"query"},{"description":"Количество элементов на странице.","required":false,"schema":{"title":"Page[Size]","exclusiveMinimum":0.0,"type":"integer","description":"Количество элементов на странице.","default":25,"lte":50},"name":"page[size]","in":"query"},{"description":"Текст полнотекстового поиска.","required":false,"schema":{"title":"Query","type":"string","description":"Текст полнотекстового поиска."},"name":"query","in":"query"},{"description":"Список id's фильмов, для фильтрации.","required":false,"schema":{"title":"Ids","type":"array","items":{"type":"string","format":"uuid"},"description":"Список id's фильмов, для фильтрации."},"name":"ids","in":"query"},{"description":"Фильтр по id жанра.","required":false,"schema":{"title":"Filter[Genre]","type":"string","description":"Фильтр по id жанра.","format":"uuid"},"name":"filter[genre]","in":"query"},{"description":"Фильтр по названию жанра.","required":false,"schema":{"title":"Filter[Genre.Name]","type":"string","description":"Фильтр по названию жанра."},"name":"filter[genre.name]","in":"query"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"film not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"films not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Список фильмов.","content":{"application/json":{"schema":{"title":"Response Film List Search Api V1 Films Search Get","type":"array","items":{"$ref":"#/components/schemas/Film"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/{film_id}":{"get":{"tags":["films"],"summary":"Информация о фильме.","description":"Детальная информация о фильме.","operationId":"film_detail_api_v1_films__film_id__get","parameters":[{"required":true,"schema":{"title":"Film Id","type":"string","format":"uuid"},"name":"film_id","in":"path"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"film not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"films not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Фильм с информацией по всем имеющимися полям.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres":{"get":{"tags":["genres"],"summary":"Главная страница жанров.","description":"На ней выводится полный список жанров.","operationId":"genre_list_api_v1_genres_get","parameters":[{"required":false,"schema":{"title":"Not Found","type":"string","default":"genre not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"genres not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Список жанров.","content":{"application/json":{"schema":{"title":"Response Genre List Api V1 Genres Get","type":"array","items":{"$ref":"#/components/schemas/Genre"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/{genre_id}":{"get":{"tags":["genres"],"summary":"Информация о жанре.","description":"Детальная информация о жанре.","operationId":"genre_detail_api_v1_genres__genre_id__get","parameters":[{"required":true,"schema":{"title":"Genre Id","type":"string","format":"uuid"},"name":"genre_id","in":"path"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"genre not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"genres not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Жанр со всеми полями.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/search":{"get":{"tags":["persons"],"summary":"Полнотекстовый поиск по персонам.","description":"Полнотекстовый поиск по персонам, по указанному параметру query в запросе.","operationId":"film_list_search_api_v1_persons_search_get","parameters":[{"description":"Сортировка по указанному полю.","required":false,"schema":{"title":"Sort","anyOf":[{"$ref":"#/components/schemas/imbdbOrderEnum"},{"$ref":"#/components/schemas/idPersonOrderEnum"}],"description":"Сортировка по указанному полю."},"name":"sort","in":"query"},{"description":"Номер страницы.","required":false,"schema":{"title":"Page[Number]","type":"integer","description":"Номер страницы.","default":0,"gte":0},"name":"page[number]","in":"query"},{"description":"Количество элементов на странице.","required":false,"schema":{"title":"Page[Size]","exclusiveMinimum":0.0,"type":"integer","description":"Количество элементов на странице.","default":25,"lte":50},"name":"page[size]","in":"query"},{"description":"Текст полнотекстового поиска.","required":false,"schema":{"title":"Query","type":"string","description":"Текст полнотекстового поиска."},"name":"query","in":"query"},{"description":"Список id's фильмов, для фильтрации.","required":false,"schema":{"title":"Ids","type":"array","items":{"type":"string","format":"uuid"},"description":"Список id's фильмов, для фильтрации."},"name":"ids","in":"query"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"person not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"persons not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Список персон.","content":{"application/json":{"schema":{"title":"Response Film List Search Api V1 Persons Search Get","type":"array","items":{"$ref":"#/components/schemas/PersonDetail"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{person_id}":{"get":{"tags":["persons"],"summary":"Информация о персоне.","description":"Детальная информация о персоне.","operationId":"person_detail_api_v1_persons__person_id__get","parameters":[{"required":true,"schema":{"title":"Person Id","type":"string","format":"uuid"},"name":"person_id","in":"path"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"person not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"persons not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Информация со всеми имеющимися полями о персоне.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{person_id}/film":{"get":{"tags":["persons"],"summary":"Информация о фильмах.","description":"Детальная информация о фильмах связанных с персоной.","operationId":"person_film_api_v1_persons__person_id__film_get","parameters":[{"required":true,"schema":{"title":"Person Id","type":"string","format":"uuid"},"name":"person_id","in":"path"},{"required":false,"schema":{"title":"Not Found","type":"string","default":"person not found"},"name":"not_found","in":"query"},{"required":false,"schema":{"title":"Not Founds","type":"string","default":"persons not found"},"name":"not_founds","in":"query"}],"responses":{"200":{"description":"Фильм с информацией по всем имеющимися полям или сама персона.","content":{"application/json":{"schema":{"title":"Response Person Film Api V1 Persons  Person Id  Film Get","type":"array","items":{"$ref":"#/components/schemas/Film"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Film":{"title":"Film","required":["uuid","title"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"imdb_rating":{"title":"Imdb Rating","type":"number"}}},"FilmDetail":{"title":"FilmDetail","required":["uuid","title"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"imdb_rating":{"title":"Imdb Rating","type":"number"},"description":{"title":"Description","type":"string","default":""},"actor_names":{"title":"Actor Names","type":"array","items":{"type":"string"},"default":[]},"writer_names":{"title":"Writer Names","type":"array","items":{"type":"string"},"default":[]},"genre":{"title":"Genre","type":"array","items":{"$ref":"#/components/schemas/Genre"},"default":[]},"actors":{"title":"Actors","type":"array","items":{"$ref":"#/components/schemas/Person"},"default":[]},"writers":{"title":"Writers","type":"array","items":{"$ref":"#/components/schemas/Person"},"default":[]},"directors":{"title":"Directors","type":"array","items":{"$ref":"#/components/schemas/Person"},"default":[]}}},"Genre":{"title":"Genre","required":["uuid","name"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"}}},"GenreDetail":{"title":"GenreDetail","required":["uuid","name"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string","default":""}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Person":{"title":"Person","required":["uuid","full_name"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"full_name":{"title":"Full Name","type":"string"}}},"PersonDetail":{"title":"PersonDetail","required":["uuid","full_name","role"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"full_name":{"title":"Full Name","type":"string"},"role":{"title":"Role","type":"string"},"film_ids":{"title":"Film Ids","type":"array","items":{"type":"string","format":"uuid"},"default":[]}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"idPersonOrderEnum":{"title":"idPersonOrderEnum","enum":["id","-id"],"type":"string","description":"An enumeration."},"imbdbOrderEnum":{"title":"imbdbOrderEnum","enum":["imdb_rating","-imdb_rating"],"type":"string","description":"An enumeration."}}}}